# Polkadot people nits and weights