--- a/chain-spec-generator/src/system_parachains_specs.rs
+++ b/chain-spec-generator/src/system_parachains_specs.rs
@@ -50,6 +50,8 @@ pub type CoretimeKusamaChainSpec = sc_chain_spec::GenericChainSpec<(), Extension
 
 pub type PeopleKusamaChainSpec = sc_chain_spec::GenericChainSpec<(), Extensions>;
 
+pub type PeoplePolkadotChainSpec = sc_chain_spec::GenericChainSpec<(), Extensions>;
+
 const ASSET_HUB_POLKADOT_ED: Balance = asset_hub_polkadot_runtime::ExistentialDeposit::get();
 
 const ASSET_HUB_KUSAMA_ED: Balance = asset_hub_kusama_runtime::ExistentialDeposit::get();
@@ -65,6 +67,8 @@ const ENCOINTER_KUSAMA_ED: Balance = encointer_kusama_runtime::ExistentialDeposi
 const CORETIME_KUSAMA_ED: Balance = coretime_kusama_runtime::ExistentialDeposit::get();
 
 const PEOPLE_KUSAMA_ED: Balance = people_kusama_runtime::ExistentialDeposit::get();
+
+const PEOPLE_POLKADOT_ED: Balance = people_polkadot_runtime::ExistentialDeposit::get();
 
 /// The default XCM version to set in genesis config.
 const SAFE_XCM_VERSION: u32 = xcm::prelude::XCM_VERSION;
@@ -142,6 +146,13 @@ pub fn bridge_hub_kusama_session_keys(keys: AuraId) -> bridge_hub_kusama_runtime
 /// The input must be a tuple of individual keys (a single arg for now since we have just one key).
 pub fn people_kusama_session_keys(keys: AuraId) -> people_kusama_runtime::SessionKeys {
 	people_kusama_runtime::SessionKeys { aura: keys }
+}
+
+/// Generate the session keys from individual elements.
+///
+/// The input must be a tuple of individual keys (a single arg for now since we have just one key).
+pub fn people_polkadot_session_keys(keys: AuraId) -> people_polkadot_runtime::SessionKeys {
+	people_polkadot_runtime::SessionKeys { aura: keys }
 }
 
 // AssetHubPolkadot
@@ -753,6 +764,80 @@ pub fn people_kusama_local_testnet_config() -> Result<Box<dyn ChainSpec>, String
 		.with_id("people-kusama-local")
 		.with_chain_type(ChainType::Local)
 		.with_genesis_config_patch(people_kusama_local_genesis(1004.into()))
+		.with_properties(properties)
+		.build(),
+	))
+}
+
+// PeoplePolkadot
+fn people_polkadot_genesis(
+	invulnerables: Vec<(AccountId, AuraId)>,
+	endowed_accounts: Vec<AccountId>,
+	id: ParaId,
+) -> serde_json::Value {
+	serde_json::json!({
+		"balances": people_polkadot_runtime::BalancesConfig {
+			balances: endowed_accounts
+				.iter()
+				.cloned()
+				.map(|k| (k, PEOPLE_POLKADOT_ED * 4096 * 4096))
+				.collect(),
+		},
+		"parachainInfo": people_polkadot_runtime::ParachainInfoConfig {
+			parachain_id: id,
+			..Default::default()
+		},
+		"collatorSelection": people_polkadot_runtime::CollatorSelectionConfig {
+			invulnerables: invulnerables.iter().cloned().map(|(acc, _)| acc).collect(),
+			candidacy_bond: PEOPLE_POLKADOT_ED * 16,
+			..Default::default()
+		},
+		"session": people_polkadot_runtime::SessionConfig {
+			keys: invulnerables
+				.into_iter()
+				.map(|(acc, aura)| {
+					(
+						acc.clone(),                         // account id
+						acc,                                 // validator id
+						people_polkadot_session_keys(aura), // session keys
+					)
+				})
+				.collect(),
+		},
+		"polkadotXcm": {
+			"safeXcmVersion": Some(SAFE_XCM_VERSION),
+		},
+		// no need to pass anything to aura, in fact it will panic if we do. Session will take care
+		// of this. `aura: Default::default()`
+	})
+}
+
+fn people_polkadot_local_genesis(para_id: ParaId) -> serde_json::Value {
+	crate::system_parachains_specs::people_polkadot_genesis(
+		// initial collators.
+		invulnerables(),
+		testnet_accounts(),
+		para_id,
+	)
+}
+
+pub fn people_polkadot_local_testnet_config() -> Result<Box<dyn ChainSpec>, String> {
+	let mut properties = sc_chain_spec::Properties::new();
+	properties.insert("ss58Format".into(), 0.into());
+	properties.insert("tokenSymbol".into(), "DOT".into());
+	properties.insert("tokenDecimals".into(), 10.into());
+
+	Ok(Box::new(
+		PeoplePolkadotChainSpec::builder(
+			people_polkadot_runtime::WASM_BINARY.expect("PeoplePolkadot wasm not available!"),
+			Extensions { relay_chain: "polkadot-local".into(), para_id: 1004 },
+		)
+		.with_name("Polkadot People Local")
+		.with_id("people-polkadot-local")
+		.with_chain_type(ChainType::Local)
+		.with_genesis_config_patch(crate::system_parachains_specs::people_polkadot_local_genesis(
+			1004.into(),
+		))
 		.with_properties(properties)
 		.build(),
 	))