Relates to: https://github.com/polkadot-fellows/runtimes/pull/319

<!-- Remember that you can run `/merge` to enable auto-merge in the PR -->

<!-- Remember to modify the changelog. If you don't need to modify it, you can check the following box.
Instead, if you have already modified it, simply delete the following line. -->

- [X] Does not require a CHANGELOG entry
- [x] generate fresh weights

### people-kusama
```
subweight compare commits          --path-pattern "./system-parachains/people/people-kusama/**/weights/**/*.rs"          --format markdown --no-color           --change added changed          --method asymptotic --ignore-errors --threshold 0.1          remotes/joepetrowski/polkadot-people origin/polkadot-people
```

(--threshold 0.1)

| File                                                                                     | Extrinsic                         | Old      | New          | Change [%] |
|------------------------------------------------------------------------------------------|-----------------------------------|----------|--------------|------------|
| system-parachains/people/people-kusama/src/weights/cumulus_pallet_parachain_system.rs    | enqueue_inbound_downward_messages | 24.42ms  | 293.81ms     | +1102.92   |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | service_page_item                 | 59.96us  | 360.34us     | +500.94    |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | execute_overweight_page_updated   | 323.67us | 360.08us     | +11.25     |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | reap_page                         | 291.37us | 305.89us     | +4.99      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | execute_overweight_page_removed   | 310.54us | 324.86us     | +4.61      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | bump_service_head                 | 157.20us | 158.49us     | +0.82      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | ready_ring_knit                   | 288.67us | 290.41us     | +0.60      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | ready_ring_unknit                 | 386.11us | 388.11us     | +0.52      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | service_page_base_no_completion   | 132.56us | 132.19us     | -0.28      |
| system-parachains/people/people-kusama/src/weights/pallet_message_queue.rs               | service_queue_base                | 129.92us | 129.47us     | -0.35      |
| system-parachains/people/people-kusama/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs | reserve_asset_deposited           |          | 18446744.07s | Added      |


## people-polkadot

```
subweight compare commits          --path-pattern "./system-parachains/people/people-polkadot/**/weights/**/*.rs"          --format markdown --no-color           --change added changed          --method asymptotic --ignore-errors --threshold 3          remotes/joepetrowski/polkadot-people origin/polkadot-people
```
(--threshold 3)

| File                                                                                       | Extrinsic                         | Old          | New          | Change [%] |
|--------------------------------------------------------------------------------------------|-----------------------------------|--------------|--------------|------------|
| system-parachains/people/people-polkadot/src/weights/cumulus_pallet_parachain_system.rs    | enqueue_inbound_downward_messages | 24.42ms      | 282.63ms     | +1057.14   |
| system-parachains/people/people-polkadot/src/weights/pallet_message_queue.rs               | service_page_item                 | 59.96us      | 357.87us     | +496.82    |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | burn_asset                        | 1.60us       | 2.20us       | +37.76     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | buy_execution                     | 1.27us       | 1.74us       | +37.01     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | clear_transact_status             | 1.26us       | 1.70us       | +34.92     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | refund_surplus                    | 1.91us       | 2.57us       | +34.48     |
| system-parachains/people/people-polkadot/src/weights/pallet_timestamp.rs                   | on_finalize                       | 2.84us       | 3.81us       | +34.06     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | expect_asset                      | 1.36us       | 1.82us       | +34.02     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | trap                              | 1.22us       | 1.63us       | +33.83     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | expect_origin                     | 1.23us       | 1.62us       | +32.24     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | expect_error                      | 1.23us       | 1.62us       | +32.14     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | set_topic                         | 1.25us       | 1.63us       | +30.82     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | expect_transact_status            | 1.47us       | 1.91us       | +29.58     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | clear_error                       | 1.24us       | 1.60us       | +29.55     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | unpaid_execution                  | 1.24us       | 1.61us       | +29.53     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | clear_topic                       | 1.24us       | 1.59us       | +28.64     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | clear_origin                      | 1.25us       | 1.61us       | +28.39     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs | receive_teleported_asset          | 2.94us       | 3.76us       | +27.89     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | set_error_handler                 | 1.30us       | 1.65us       | +26.92     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | set_appendix                      | 1.32us       | 1.65us       | +25.28     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | set_fees_mode                     | 1.30us       | 1.61us       | +24.13     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | descend_origin                    | 1.42us       | 1.76us       | +24.12     |
| system-parachains/people/people-polkadot/src/weights/cumulus_pallet_xcmp_queue.rs          | take_first_concatenated_xcm       | 6.12us       | 7.50us       | +22.47     |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | expect_pallet                     | 3.77us       | 4.60us       | +22.05     |
| system-parachains/people/people-polkadot/src/weights/pallet_utility.rs                     | as_derivative                     | 3.95us       | 4.81us       | +21.68     |
| system-parachains/people/people-polkadot/src/weights/pallet_utility.rs                     | dispatch_as                       | 5.92us       | 6.99us       | +18.03     |
| system-parachains/people/people-polkadot/src/weights/pallet_utility.rs                     | force_batch                       | 2.79ms       | 3.29ms       | +17.89     |
| system-parachains/people/people-polkadot/src/weights/pallet_utility.rs                     | batch                             | 2.79ms       | 3.28ms       | +17.52     |
| system-parachains/people/people-polkadot/src/weights/pallet_utility.rs                     | batch_all                         | 2.98ms       | 3.50ms       | +17.44     |
| system-parachains/people/people-polkadot/src/weights/frame_system.rs                       | set_code                          | 110.07ms     | 123.55ms     | +12.25     |
| system-parachains/people/people-polkadot/src/weights/pallet_message_queue.rs               | execute_overweight_page_updated   | 323.67us     | 357.98us     | +10.60     |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | proxy                             | 38.95us      | 42.01us      | +7.87      |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | transact                          | 7.74us       | 8.32us       | +7.48      |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_generic.rs  | query_response                    | 31.55us      | 33.41us      | +5.90      |
| system-parachains/people/people-polkadot/src/weights/pallet_xcm.rs                         | claim_assets                      | 158.82us     | 168.06us     | +5.82      |
| system-parachains/people/people-polkadot/src/weights/frame_system.rs                       | apply_authorized_upgrade          | 113.43ms     | 119.88ms     | +5.69      |
| system-parachains/people/people-polkadot/src/weights/pallet_message_queue.rs               | reap_page                         | 291.37us     | 306.29us     | +5.12      |
| system-parachains/people/people-polkadot/src/weights/pallet_xcm.rs                         | take_response                     | 150.21us     | 157.65us     | +4.95      |
| system-parachains/people/people-polkadot/src/weights/pallet_message_queue.rs               | execute_overweight_page_removed   | 310.54us     | 325.82us     | +4.92      |
| system-parachains/people/people-polkadot/src/weights/pallet_balances.rs                    | transfer_allow_death              | 169.03us     | 176.56us     | +4.45      |
| system-parachains/people/people-polkadot/src/weights/pallet_identity.rs                    | cancel_request                    | 150.80us     | 157.32us     | +4.33      |
| system-parachains/people/people-polkadot/src/weights/pallet_balances.rs                    | transfer_all                      | 168.54us     | 175.76us     | +4.28      |
| system-parachains/people/people-polkadot/src/weights/pallet_balances.rs                    | force_adjust_total_issuance       | 30.57us      | 31.84us      | +4.16      |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | remove_proxy                      | 146.73us     | 152.82us     | +4.15      |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | create_pure                       | 146.77us     | 152.79us     | +4.10      |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | add_proxy                         | 146.86us     | 152.50us     | +3.84      |
| system-parachains/people/people-polkadot/src/weights/pallet_xcm.rs                         | teleport_assets                   | 407.90us     | 423.25us     | +3.76      |
| system-parachains/people/people-polkadot/src/weights/pallet_identity.rs                    | request_judgement                 | 178.87us     | 185.39us     | +3.64      |
| system-parachains/people/people-polkadot/src/weights/pallet_balances.rs                    | transfer_keep_alive               | 159.77us     | 165.51us     | +3.59      |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs | withdraw_asset                    | 154.16us     | 159.11us     | +3.21      |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | kill_pure                         | 146.16us     | 150.82us     | +3.19      |
| system-parachains/people/people-polkadot/src/weights/pallet_xcm.rs                         | already_notified_target           | 168.46us     | 173.83us     | +3.19      |
| system-parachains/people/people-polkadot/src/weights/pallet_proxy.rs                       | remove_proxies                    | 145.43us     | 149.92us     | +3.09      |
| system-parachains/people/people-polkadot/src/weights/pallet_collator_selection.rs          | add_invulnerable                  | 467.79us     | 452.76us     | -3.21      |
| system-parachains/people/people-polkadot/src/weights/pallet_collator_selection.rs          | leave_intent                      | 303.17us     | 291.62us     | -3.81      |
| system-parachains/people/people-polkadot/src/weights/pallet_collator_selection.rs          | update_bond                       | 198.15us     | 187.85us     | -5.20      |
| system-parachains/people/people-polkadot/src/weights/pallet_identity.rs                    | remove_expired_approval           | 150.90us     | 140.29us     | -7.03      |
| system-parachains/people/people-polkadot/src/weights/frame_system.rs                       | remark_with_event                 | 7.20ms       | 6.67ms       | -7.41      |
| system-parachains/people/people-polkadot/src/weights/pallet_multisig.rs                    | as_multi_threshold_1              | 20.04us      | 18.29us      | -8.73      |
| system-parachains/people/people-polkadot/src/weights/frame_system.rs                       | remark                            | 1.59ms       | 1.39ms       | -12.60     |
| system-parachains/people/people-polkadot/src/weights/pallet_xcm.rs                         | execute                           | 18446744.07s | 8.91us       | -100.00    |
| system-parachains/people/people-polkadot/src/weights/xcm/pallet_xcm_benchmarks_fungible.rs | reserve_asset_deposited           |              | 18446744.07s | Added      |