@acatangiu 2024-05-30
why is `enqueue_inbound_downward_messages` taking 300ms now vs 25ms before?

is it that big for other parachains too?