@bkontur 2024-05-30
> why is `enqueue_inbound_downward_messages` taking 300ms now vs 25ms before?
> 
> is it that big for other parachains too?

@acatangiu good question, when I check the big PR with fresh weights for 1.2.0 release: https://github.com/polkadot-fellows/runtimes/pull/223, everywhere it was around 300ms. I think that those weights for `enqueue_inbound_downward_messages` are just copied from [polkadot-sdk](https://github.com/paritytech/polkadot-sdk/blob/master/cumulus/parachains/runtimes/people/people-rococo/src/weights/cumulus_pallet_parachain_system.rs#L16) where is not command and just `//! Need to rerun`